requirements.txt
google-cloud-storage>=2.0.0
google-cloud-iam>=2.0.0
google-cloud-resource-manager>=1.0.0
python-dotenv>=0.19.0